mcp/technical_analyst/pyproject.toml
[project]
name = "technical-analyst-mcp"
version = "0.1.0"
description = "Alpha-Guardian Technical Analyst MCP Server"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "yfinance>=0.2.36",
    "pandas>=2.2.0",
    "pandas_ta",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
]

[tool.uv]
package = true

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false
